+++ assets/js/models/projectContributors/index.tsx
@@ -4,8 +4,8 @@ import { match } from "ts-pattern";
 export type ProjectContributor = api.ProjectContributor;
 export type ContributorRole = "champion" | "reviewer" | "contributor";
 
-export const CHAMPION_RESPONSIBILITY = "Champion - Responsible for the success of the project";
-export const REVIEWER_RESPONSIBILITY = "Reviewer - Responsible for reviewing and acknowledging progress";
+const CHAMPION_RESPONSIBILITY = "Champion - Responsible for the success of the project";
+const REVIEWER_RESPONSIBILITY = "Reviewer - Responsible for reviewing and acknowledging progress";
 
 export function responsibility(contributor: ProjectContributor | undefined, role: ContributorRole) {
   return match(role)
