+++ assets/js/models/projects/index.tsx
@@ -27,8 +27,6 @@ export const PRIVACY_INTERNAL = "internal";
 export const PRIVACY_CONFIDENTIAL = "confidential";
 export const PRIVACY_SECRET = "secret";
 
-export { groupBySpace } from "./groupBySpace";
-
 export function useProjectContributorCandidates(id: string): (query: string) => Promise<api.Person[]> {
   return async (query: string) => {
     return await api.searchProjectContributorCandidates({ projectId: id, query: query }).then((r) => r.people!);
@@ -68,10 +66,6 @@ export function isMilestoneOverdue(milestone: Pick<Milestone, "status" | "deadli
   if (!day) return false;
 
   return !Time.isToday(day) && Time.isPast(day);
-}
-
-export function allMilestonesCompleted(project: Project) {
-  return project.milestones!.every((m) => m!.status === "done");
 }
 
 export function isPausable(project: Project) {
