--- assets/js/pages/AccountPage/loader.tsx
@@ -1,3 +1,0 @@
-export async function loader() {
-  return null;
-}
