+++ assets/js/pages/ProjectEditResourcesPage/loader.tsx
@@ -19,7 +19,3 @@ export async function loader({ params }): Promise<LoaderResult> {
 export function useLoadedData(): LoaderResult {
   return Pages.useLoadedData() as LoaderResult;
 }
-
-export function useRefresh() {
-  return Pages.useRefresh();
-}
